+++ Nubimetrics/Nubimetrics.Composition/DomainModule.cs
@@ -1,7 +1,9 @@
 ﻿using Autofac;
 using Nubimetrics.Domain.Country;
+using Nubimetrics.Domain.Currency;
 using Nubimetrics.Domain.SearchItem;
 using Nubimetrics.DomainContracts.Country;
+using Nubimetrics.DomainContracts.Currency;
 using Nubimetrics.DomainContracts.SearchItem;
 
 namespace Nubimetrics.Composition
@@ -12,6 +14,7 @@ namespace Nubimetrics.Composition
         {
             builder.RegisterType<CountryConfiguration>().As<ICountryConfiguration>();
             builder.RegisterType<SearchItemConfiguration>().As<ISearchItemConfiguration>();
+            builder.RegisterType<CurrencyConfiguration>().As<ICurrencyConfiguration>();
         }
     }
 }
