+++ Nubimetrics/Nubimetrics.Composition/MappingBuilder.cs
@@ -10,7 +10,7 @@ namespace Nubimetrics.Composition
     {
         public MappingBuilder()
         {
-            CreateMap<Item, ItemDto>()
+            CreateMap<ItemDto, Item>()
                 .ForMember(dest => dest.site_id, opt => opt.MapFrom(src => src.site_id))
                 .ForMember(dest => dest.country_default_time_zone, opt => opt.MapFrom(src => src.country_default_time_zone))
                 .ForMember(dest => dest.query, opt => opt.MapFrom(src => src.query))
@@ -20,12 +20,12 @@ namespace Nubimetrics.Composition
                 .ForMember(dest => dest.available_sorts, opt => opt.MapFrom(src => src.available_sorts))
                 .ForMember(dest => dest.filters, opt => opt.MapFrom(src => src.filters))
                 .ForMember(dest => dest.available_filters, opt => opt.MapFrom(src => src.available_filters));
-            CreateMap<Result, ResultDto>()
+            CreateMap<ResultDto, Result>()
                 .ForMember(dest => dest.id, opt => opt.MapFrom(src => src.id))
                 .ForMember(dest => dest.site_id, opt => opt.MapFrom(src => src.site_id))
                 .ForMember(dest => dest.title, opt => opt.MapFrom(src => src.title))
                 .ForMember(dest => dest.price, opt => opt.MapFrom(src => src.price))
-                .ForMember(dest => dest.seller, opt => opt.MapFrom(src => src.seller))
+                .ForMember(dest => dest.seller_Id, opt => opt.MapFrom(src => src.seller.id))
                 .ForMember(dest => dest.permalink, opt => opt.MapFrom(src => src.permalink));
         }
     }
