+++ Nubimetrics/Nubimetrics.DalContracts/Meli/IMeliWebRepository.cs
@@ -1,8 +1,7 @@
 ﻿using Nubimetrics.DalEntities.Country;
+using Nubimetrics.DalEntities.Currency;
 using Nubimetrics.DalEntities.SearchItem;
-using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Nubimetrics.DalContracts.Meli
@@ -10,6 +9,9 @@ namespace Nubimetrics.DalContracts.Meli
     public interface IMeliWebRepository
     {
         Task<CountryDto> GetCountryAsync(string country);
-        Task<Item> GetItemAsync(string item);
+        Task<ItemDto> GetItemAsync(string item);
+        Task<List<CurrencyDto>> GetCurrencyAsync();
+        Task<CurrencyConvertionDto> GetCurrencyConvertionAsync(string currency);
+
     }
 }
