+++ Nubimetrics/Nubimetrics.DomainContracts/Currency/ICurrencyConfiguration.cs
@@ -1,0 +1,11 @@
+﻿using Nubimetrics.DalEntities.Currency;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace Nubimetrics.DomainContracts.Currency
+{
+    public interface ICurrencyConfiguration
+    {
+        Task<List<CurrencyDto>> GetCurrencyAsync(string path, string fileName, string extension);
+    }
+}
