+++ Nubimetrics/Nubimetrics.Entities/Currency/CurrencyDto.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Nubimetrics.DalEntities.Currency
+{
+    public class CurrencyDto
+    {
+        public string id { get; set; }
+        public string symbol { get; set; }
+        public string description { get; set; }
+        public int decimal_places { get; set; }
+        public CurrencyConvertionDto ToDolar { get; set; }
+    }
+}
