+++ Nubimetrics/Nubimetrics.Repository/Repositories/MeliWebRepository.cs
@@ -1,13 +1,9 @@
-﻿using AutoMapper;
-using Microsoft.IdentityModel.Clients.ActiveDirectory;
-using Nubimetrics.DalContracts.Meli;
+﻿using Nubimetrics.DalContracts.Meli;
 using Nubimetrics.DalEntities.Country;
+using Nubimetrics.DalEntities.Currency;
 using Nubimetrics.DalEntities.SearchItem;
-using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Nubimetrics.Dal.Repositories.Meli
@@ -18,10 +14,17 @@ namespace Nubimetrics.Dal.Repositories.Meli
         {
            return await GetDataAsync<CountryDto>(string.Format(ApiResource.MeliCountries, country), HttpMethod.Get);        
         }
-
-        public async Task<Item> GetItemAsync(string item)
+        public async Task<ItemDto> GetItemAsync(string item)
         {
-            return await GetDataAsync<Item>(string.Format(ApiResource.MeliItem, item), HttpMethod.Get);            
+            return await GetDataAsync<ItemDto>(string.Format(ApiResource.MeliItem, item), HttpMethod.Get);            
+        }
+        public async Task<List<CurrencyDto>> GetCurrencyAsync()
+        {
+            return await GetDataAsync<List<CurrencyDto>>(string.Format(ApiResource.MeliCurrency), HttpMethod.Get);
+        }
+        public async Task<CurrencyConvertionDto> GetCurrencyConvertionAsync(string currency)
+        {
+            return await GetDataAsync<CurrencyConvertionDto>(string.Format(ApiResource.MeliCurrencyToDolar, currency), HttpMethod.Get);
         }
     }
 }
