+++ Nubimetrics/Nubimetrics/Nubimetrics.Api.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="7.2.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
     <PackageReference Include="System.Text.Json" Version="6.0.1" />
   </ItemGroup>
